shodan_il_exposure_starter/requirements.txt
shodan
pandas
matplotlib
python-dotenv